requirements.txt
streamlit
psycopg2-binary
pandas
plotly
python-dotenv
requests
matplotlib
chronos-forecasting
torch
accelerate
